Close JPEG-XL decoding/screenshot support, #3852
This commit will:

- Remove ppm, pgm, pgmyuv and tga from the Preferences.ScreenshotFormat
  enum
- Add webp and jxl to the Preferences.ScreenshotFormat enum
- Add WebP (.webp) and JPEG XL (.jxl) as choices for the Format setting
  in the Screenshots section of General settings
- Add a new method createNSImageWithContentsOf to FFmpegController
- Add a new private method createImage to PlayerCore that will call the
  createNSImageWithContentsOf method if the normal NSImage initializer
  is unable to decode the screenshot
- Change PlayerCore.screenshotCallback to use the new method to create a
  NSImage to preview in the OSD
- Add a new class FFmpegLogger that provides access to the normal IINA
  logger to the FFmpegController
- Change the FFmpegController to use the IINA logger instead of NSLog
  when logging messages
- Add an internal setting enableFFmpegImageDecoder to allow disabling
  use of the FFmpeg image decoder should a problem be uncovered
- Change the screenshot preview animation duration to adhere to
  accessibility settings

NOTE that removing enumeration cases breaks backward compatibility as
the new cases reuse the tag values previously assigned to ppm and pgm.
These four screenshot formats were dropped by mpv and removed from
IINA's setting panel back in 2017. Anyone using these old formats would
have had to update their settings therefore it should be acceptable to
reuse the tag values.

The main author of the settings changes is Carter Li. That portion of
the PR is a merge of changes that have been tested in his IINA fork.

The FFmpeg related changes are new and support the ability to show
screenshot previews in the OSD for image formats not supported by
AppKit. Currently JPEG XL is not supported by the NSImage initializer.
This is also true for WebP in macOS versions before Big Sur.

Co-authored-by: CarterLi <carter.li@eoitek.com>